+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fixed an error that could occur when querying products via GraphQL.
 - Fixed an error that could occur when querying variants via GraphQL.
 - Fixed a bug where untracked stock items did not show correctly on the example templates. ([#3510](https://github.com/craftcms/commerce/issues/3510))
+- Fixed a bug where disabling a pricing rule wouldn’t update catalog prices. ([#3515](https://github.com/craftcms/commerce/issues/3515))
 - Added `craft\commerce\gql\types\input\IntFalse`.
 
 ## 5.0.6 - 2024-05-15